autoload.php.dist - fix ClassLoader
The Symfony project has [removed the `UniversalClassLoader` component](https://github.com/symfony/ClassLoader/commit/2ae376a83b03004c8427b414ba4ba1b394179889). The current autoload.php.dist is broken because of this. I had to make the attached edits to get this to work.  (I'm generally unfamiliar with autoloading and namespaces, but there doesn't appear to be an equivalent functionality in the `ClassLoader` class to the following chunk of code, so I just took it out.

```php
$loader->registerNamespaces(array(
    'Console' => BASE_PATH . '/bin',
    'Symfony' => BASE_PATH . '/vendor/symfony/src',
));
```