Removes 'target-dir' from composer.json removes unneeded files
The 'target-dir' config in the composer.json causes the repository to be
downloaded to the path specified there. This caused issues with the
autoloading. This should be fixed now. Also, the vendor directory is
removed because it is being ignored in the .gitignore file and the
nbproject folder is not needed.